requirements.txt
Flask==2.0.1
numpy==1.21.2
pandas==1.3.3
scikit-learn==1.0
requests==2.26.0
beautifulsoup4==4.10.0
tldextract==3.1.2
python-whois==0.7.3
googlesearch-python==2020.0.2